+++ doc.go
@@ -1,0 +1,2 @@
+// Package pinyin : 汉语拼音转换工具
+package pinyin