+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.2.0 (2015-01-04)
+
+* 新增 ``func NewArgs() Args``
+* 解决 ``Args.Separator`` 无法赋值为 ``""`` 的 BUG
+* 规范命名:
+    * ``NORMAL`` -> ``Normal``
+    * ``TONE`` -> ``Tone``
+    * ``TONE2`` -> ``Tone2``
+    * ``INITIALS`` -> ``Initials``
+    * ``FIRST_LETTER`` -> ``FirstLetter``
+    * ``FINALS`` -> ``Finals``
+    * ``FINALS_TONE`` -> ``FinalsTone``
+    * ``FINALS_TONE2`` -> ``FinalsTone2``
+
 ## 0.1.1 (2014-12-07)
 * 更新拼音库